Split D3Chart utils axis file
